requirements.txt
opencv-python
face-recognition
numpy
pillow
tk
pandas
fpdf
